refactoring components folder
